GC2D/GCConsole2 drawJuice: 86.8% -> 87.1% (use byte pane index)

+++ src/GC2D/GCConsole2.cpp
@@ -4006,7 +4006,7 @@ void TGCConsole2::drawJuice(J2DOrthoGraph& graph, u32 color)
 	loadPictureTexture((J2DPicture*)unk328, GX_TEXMAP1);
 
 	int selected = 0x17;
-	for (int i = 0x16; (u8)i != 0; --i) {
+	for (u8 i = 0x16; i != 0; --i) {
 		int threshold = (int)(64.0f * (1.0f - unk2B8));
 		if (unk334[i]->mBounds.y1 >= threshold) {
 			selected = i + 1;
